added authorization gates to user, role , product and order controllers

+++ app/Http/Controllers/ProductController.php
@@ -1,11 +1,12 @@
 <?php
 namespace App\Http\Controllers;
 
-use App\Http\Requests\ProductCreateRequest;
-use Illuminate\Support\Facades\Storage;
 use App\Product;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Resources\ProductResource;
+use Illuminate\Support\Facades\Storage;
+use App\Http\Requests\ProductCreateRequest;
 use Symfony\Component\HttpFoundation\Response;
 
 class ProductController extends Controller
@@ -17,6 +18,7 @@ class ProductController extends Controller
      */
     public function index()
     {
+        Gate::authorize('view', 'products');
         $products = Product::paginate(5);
         return ProductResource::collection($products);
     }
@@ -29,6 +31,8 @@ class ProductController extends Controller
      */
     public function store(ProductCreateRequest $request)
     {
+        Gate::authorize('edit', 'products');
+
         $product = Product::create($request->only('title', 'description', 'image', 'price'));
 
         return response($product, Response::HTTP_CREATED);
@@ -42,6 +46,8 @@ class ProductController extends Controller
      */
     public function show(Product $product)
     {
+        Gate::authorize('view', 'products');
+
         return new ProductResource($product);
     }
 
@@ -54,6 +60,8 @@ class ProductController extends Controller
      */
     public function update(Request $request, Product $product)
     {
+        Gate::authorize('edit', 'products');
+
         $product->update($request->only('title', 'description', 'image', 'price'));
 
         return response($product, Response::HTTP_CREATED);
@@ -67,6 +75,8 @@ class ProductController extends Controller
      */
     public function destroy(Product $product)
     {
+        Gate::authorize('edit', 'products');
+
         Product::destroy($product->id);
         return response(null, Response::HTTP_NO_CONTENT);
     }
